fix/#245 : 버전체크 로직 이동
