+++ app/src/main/java/com/depromeet/team6/presentation/ui/home/HomeContract.kt
@@ -82,6 +82,9 @@ class HomeContract {
         data class NavigateToItinerary(val markerParameter: FocusedMarkerParameter?) : HomeSideEffect
         data object ShowDeleteAlarmToast : HomeSideEffect
         data class ShowToast(val message: String) : HomeSideEffect
+
+        data object ShowUpdateRequiredDialog : HomeSideEffect
+        data object ShowUpdateOptionalDialog : HomeSideEffect
     }
 
     sealed class HomeEvent : UiEvent {
